python/requirements.txt
pandas>=2.0
altair>=5.0
requests>=2.28
vl-convert-python>=1.0
python-dotenv>=1.0
